replaced all usages of 'double' into 'BigDecimal' including implementation of object

+++ app/src/main/java/com/simsdroid/app/ModelOrderView.java
@@ -1,12 +1,14 @@
 package com.simsdroid.app;
 
+import java.math.BigDecimal;
+
 public class ModelOrderView {
     public long orderNumber;
     public String date;
     public String time;
-    public double total;
+    public BigDecimal total;
 
-    public ModelOrderView(long orderNumber, String date, String time, double total) {
+    public ModelOrderView(long orderNumber, String date, String time, BigDecimal total) {
         this.orderNumber = orderNumber;
         this.date = date;
         this.time = time;
